backend/pyproject.toml
[tool.poetry]
name = "interactive-portfolio"
version = "0.1.0"
description = ""
authors = ["OusmaneKK"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
alembic = "1.12.0"
apicrud = "0.3.6"
asgiref = "3.7.2"
async-timeout = "4.0.3"
attrs = "23.1.0"
authlib = "0.15.6"
b2sdk = "1.24.1"
babel = "2.13.0"
boto3 = "1.28.68"
botocore = "1.31.68"
cachetools = "5.3.1"
certifi = "2023.7.22"
cffi = "1.16.0"
charset-normalizer = "3.3.1"
click = "8.1.7"
clickclick = "20.10.2"
connexion = "2.14.2"
coverage = "7.3.4"
cryptography = "41.0.4"
decorator = "5.1.1"
defusedxml = "0.7.1"
django = "4.2.6"
django-cors-headers = "4.3.0"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "5.3.0"
flask = "2.2.5"
flask-babel = "2.0.0"
flask-cors = "4.0.0"
future = "0.18.3"
geocoder = "1.38.1"
greenlet = "3.0.0"
gunicorn = "21.2.0"
httpie = "3.2.2"
idna = "3.4"
inflection = "0.5.1"
itsdangerous = "2.1.2"
jinja2 = "3.1.2"
jmespath = "1.0.1"
jsonschema = "4.19.1"
jsonschema-specifications = "2023.7.1"
ldap3 = "2.9.1"
logfury = "1.0.1"
mako = "1.2.4"
markdown-it-py = "3.0.0"
markupsafe = "2.1.3"
mdurl = "0.1.2"
multidict = "6.0.4"
packaging = "23.2"
passlib = "1.7.4"
pillow = "10.1.0"
psycopg2 = "2.9.9"
pyasn1 = "0.5.0"
pycparser = "2.21"
pycryptodomex = "3.19.0"
pydub = "0.25.1"
pygments = "2.16.1"
pyjwt = "2.8.0"
pymysql = "1.1.0"
pyotp = "2.9.0"
pysocks = "1.7.1"
python-dateutil = "2.8.2"
python-decouple = "3.8"
pytz = "2023.3.post1"
pyyaml = "6.0.1"
ratelim = "0.1.6"
redis = "5.0.1"
referencing = "0.30.2"
requests = "2.31.0"
requests-toolbelt = "1.0.0"
rich = "13.6.0"
rpds-py = "0.10.6"
s3transfer = "0.7.0"
six = "1.16.0"
sqlalchemy = "1.3.24"
sqlalchemy-utils = "0.41.1"
sqlparse = "0.4.4"
swagger-ui-bundle = "0.0.9"
tqdm = "4.66.1"
typing-extensions = "4.8.0"
urllib3 = "2.0.7"
werkzeug = "2.2.3"
ffmpeg = "^1.4"
tox = "^4.11.4"


[tool.poetry.group.dev.dependencies]
pytest-django = "^4.7.0"
pytest = "^7.4.3"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "myapp.settings"
testpaths = [
    "myapp/tests/tests.py"
]
addopts = "-v -s --cov=apicrud --no-cov-on-fail --cov-report=term-missing"
